수정: 파편화된 유저 비지니스 로직 병합
