+++ src/main/java/com/devcourse/kurlymurly/web/common/ApiResponse.java
@@ -1,4 +1,4 @@
-package com.devcourse.kurlymurly.web.controller.user;
+package com.devcourse.kurlymurly.web.common;
 
 import org.springframework.http.HttpStatus;
