+++ src/test/java/com/devcourse/kurlymurly/module/user/domain/UserRepositoryTest.java
@@ -1,12 +1,9 @@
 package com.devcourse.kurlymurly.module.user.domain;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-
-import java.sql.Date;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -19,12 +16,12 @@ class UserRepositoryTest {
     @Test
     @DisplayName("동일한 아이디가 있는지 확인")
     void existsByLoginId() {
-        assertThat(userRepository.existsByLoginId("kurly")).isEqualTo(false);
+        assertThat(userRepository.existsByLoginId("kurly")).isFalse();
     }
 
     @Test
     @DisplayName("동일한 이메일이 있는지 확인")
     void existsByEmail() {
-        assertThat(userRepository.existsByEmail("kkk@murly.com")).isEqualTo(false);
+        assertThat(userRepository.existsByEmail("kkk@murly.com")).isFalse();
     }
 }
