+++ src/test/java/com/devcourse/kurlymurly/module/user/service/UserServiceTest.java
@@ -1,11 +1,14 @@
 package com.devcourse.kurlymurly.module.user.service;
 
+import com.devcourse.kurlymurly.module.product.service.ProductFacade;
 import com.devcourse.kurlymurly.module.user.domain.UserRepository;
+import com.devcourse.kurlymurly.module.user.domain.cart.CartRepository;
 import com.devcourse.kurlymurly.web.dto.user.JoinUser;
 import com.devcourse.kurlymurly.web.exception.ExistUserInfoException;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
@@ -13,9 +16,14 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
+import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.BDDMockito.then;
+import static org.mockito.BDDMockito.willDoNothing;
+import static org.mockito.BDDMockito.willThrow;
 import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.times;
 
 @ExtendWith(MockitoExtension.class)
 class UserServiceTest {
@@ -27,6 +35,12 @@ class UserServiceTest {
 
     @Mock
     private PasswordEncoder passwordEncoder;
+
+    @Mock
+    private ProductFacade productFacade;
+
+    @Mock
+    private CartRepository cartRepository;
 
     private static JoinUser.Request user;
 
@@ -76,5 +90,38 @@ class UserServiceTest {
 
         // Then
         assertThrows(ExistUserInfoException.class, () -> userService.join(user));
+    }
+
+    @Nested
+    @DisplayName("카트 테스트")
+    class cartTest {
+        private final Long userId = 1L;
+        private final Long productId = 1L;
+        private final int quantity = 2;
+
+        @Test
+        @DisplayName("주문 가능한 상품은 장바구니에 추가할 수 있다.")
+        void addCart_Success() {
+            // given
+            willDoNothing().given(productFacade).validateOrderable(any());
+
+            // when
+            userService.addCart(userId, productId, quantity);
+
+            // then
+            then(cartRepository).should(times(1)).save(any());
+        }
+
+        @Test
+        @DisplayName("삭제되었거나 품절인 상품을 장바구니에 담으려고 하면 IllegalStateException을 던진다.")
+        void addCart_Fail_ByInvalidStatus() {
+            // given
+            willThrow(IllegalStateException.class).given(productFacade).validateOrderable(any());
+
+            // when, then
+            assertThatExceptionOfType(IllegalStateException.class)
+                    .isThrownBy(() -> userService.addCart(userId, productId, quantity));
+            then(cartRepository).shouldHaveNoInteractions();
+        }
     }
 }
